move webserver in its own folder
